Removed verification search. Adjusted LMR.
Verification search appeared to actually worsen play. I thought this was strange, but after reading through the forums Robert Hyatt had the same issue. While my program is no where near comparable to Crafty, it seems that Verification could be a net loss of ELO

Changed the condition on LMR to be 'must have found table move'. This is probably too aggresive in some cases, and not aggresive enough in others. Additionally, it would make sense to make sure that the tableMove which was found, is infact the zeroth move in the moveList, something that I did not do in this version for testing.