Updated Formula for Null Move R
Now using formula similar to stockfish, but is clearly very linear. R starts at 2.5 and increases by .25 per Depth. Depth 2 has an R value of 3, Depth 6 has an R value of 4...

Formula is (40 + (depth << 2)) >> 4

May be some value in taking the ceil over the floot of the result depending on how much larger the eval is than beta

Tests show ~10 ELO gain after ~2000 games. More testing is needed, but this change should stay as an improvement