Added Mobility. Fixed bug.
EloTables now includes a listing for Ethereal7.65, even though 7.65 is almost identical to the previous 7.61. The Elo difference is a result of the update in the testing framework. However, some elo may be attributable to Fancy Magic BitBoards that have been added.
Added precomputed popcounts of each index of the look up tables as a means to determine mobility. In order to avoid popcount during the evaluation, we consider any possible attack, move, or defense, as a move for piece.
Updated movegen.h, evaluate.c, evaluate.h, magics.c and magics.h to make this happen.
Fixed a bug which could cause segfaults whenever the pawn hash key 0xFFFF was used. This was pointed out to me by Antonio Arias. As a result, I also identified issues with the History counters, although they would never cause a crash.

Elo gain = ~30