No longer jump into quiscence search while in check
Determine check status perform jumping into qsearch, and avoid the jump if the board is in check. Increase the depth to 1 and set a flag to avoid extending the check again.
Updated EloTables
Removed old Android compile. Will recompile for 7.70. Old version was at least 100commits back and not UCI ready