Fixed an engine crash, and an uninitalized value error
Ethereal would crash upon recieving a position command, if its length exceeded 2048 characters. I felt confident that no position would go that long, but I was wrong. The limit is now 8192 characters, which is ~2000 moves. I would be shocked to see a game go that long before adjudicating
Sometimes the LPV would not be filled during IID, causing bad values to be used as the table move. I could not find a case where fixing the issue changed the node counts, but it resolves the valgrind complaints none the less