Add support for popcount compilation
First time trying this... so will have to confirm that others
can use the compiles. Specifically, the few users who do not
have support for popcount in hardware.

Speed up based on a bench 23 of the new release versions,
note that the last commit used non-release versions, which have
a -march=native flag that is NOT set when making release files

Time to bench 23 (127078ms -> 119466ms) ~6% speedup

BENCH : 9,759,585