Combine MG and EG into one integer
SF folks have been doing this for quite some time. This makes
things a little less clean on the low level side of things, but
it makes working with the evaluation function a little nicer.

I've run bench 25's on both base and this patch. The node count
differs by a total of 7. I imagine this is some case where eval
ends up getting a bad value due to some overflow in an intermediate
step of the evaluation. I have tried to curb that by doing the eval
in steps for white and black, to balance out the scores. I expect
that the only time this overflow happens is when one player has
an absurd amount of material, which should not happen in any real
games, or very often in the search tree.

I'm going to commit anyway because 1) The problem likely already happens
and 2) We accept 64bit hash TransTable instability, why not this?

Bench 25: NPS (2.526K -> 2.617K) = 3.6% speedup

SPRT @ 4.0+0.04s Threads=1 Hash=2MB
LLR: 2.27 (-2.20, 2.20) [0.00, 4.00]
Games: 23900 W: 6232 L: 5995 D: 11673

BENCH : 7,946,591