Hash replacment policy with a focus on SMP
On 1thread it is very rare to overwrite a hash entry from
the same position in a situation where we would not want to,
as the depth is usually increasing with each update.

However, with SMP this is not the case. This patch adds the
Stockfish method (verbatim) as a final condition to replace
an existing entry. Limited test shows that this has the potential
to scale very well with more time and with more threads.

Testing is limited, and bounds are lowered, as LTC with 7threads
is an extremely expensive process to run

SPRT @ 3.0+0.03s Threads=1 Hash=1MB
LLR: 2.30 (-2.20, 2.20) [-4.00, 0.00]
Games: 19825 W: 5204 L: 5157 D: 9464

SPRT @ 15.0+0.15s Threads=1 Hash=8MB
LLR: 1.90 (-2.20, 2.20) [-4.00, 0.00]
Games: 8775 W: 1892 L: 1821 D: 5062

SPRT @ 3.0+0.03s Threads=5 Hash=64MB
LLR: 2.20 (-2.20, 2.20) [0.00, 5.00]
Games: 6975 W: 1764 L: 1626 D: 3585

SPRT @ 15.0+0.15s Threads=7 Hash=512MB
LLR: 2.23 (-2.20, 2.20) [-1.50, 4.50]
Games: 1325 W: 319 L: 234 D: 772

BENCH : 9,076,028