Added support for Syzygy through Fathom (Jon Dart's updated Fork)
I do not currently have an SSD, so testing was limited to a standard HD...

Testing with ProbeDepth=8

Ethereal 9.64: 4083 - 3751 - 11870  [0.508] 19704 @ 20+.2s and 8MB Hash

I've not yet seen any crashes running with TB. If you do, please inform me.

My hope is that the elo gain is much larger with the proper hardware.