Tune minors; Large changes to the tuner
Add back the tune-by-depth which worked well in the past, and has
proven to work again with this patch, the recent threat tuning,
as well as with one of Laldon's recent tunes.

Add the setting CLEARING to the tuner, which when set to 1 ensures
correctness from a hashing and history perspective, but is extremely
expensive. For low depth tunes, it might be worth leaving it off. For
long searches, it becomes a greater issue to have it not cleared.

ELO   | 1.76 +- 1.19 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 2.99 (-2.94, 2.94) [0.00, 4.00]
Games | N: 141500 W: 30905 L: 30188 D: 80407

ELO   | 2.48 +- 1.99 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 2.97 (-2.94, 2.94) [0.00, 4.00]
Games | N: 39600 W: 6833 L: 6550 D: 26217

BENCH : 6,645,815