Rewrite History Heuristic underlying data structures
There is a slight functional change, but only for the following case.
Before, the Best Move was the first to have it's history updated, and now
it is updated last, as in accordance with the quietsTried array.

The bench changes when there exists a move other than the best move which
uses the same piece type and has the same final destination. This difference
has been proven to be non impactful through testing.

ELO   | 2.29 +- 1.81 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 2.98 (-2.94, 2.94) [0.00, 4.00]
Games | N: 58530 W: 12341 L: 11956 D: 34233
http://chess.grantnet.us/viewTest/3162/

BENCH : 7,811,229