Fix an oversight from 11.39 causing illegal PVs
I made the mistake of not zeroing out the length of PV, when I
should have skipped zeroing the lpv, since the original PV that
is passed down the search comes from the search info and is never
zeroed, leading to illegal pv moves. No illegal moves are played.

NO FUNCTIONAL CHANGE

BENCH : 7,585,979