Adjust the Tuning method for better results
Originally, we would simply save the board eval, and try to
adjust that to match the result. Now, we will use 50% the board
eval, and 50% a depth 10 search that was generated and added to
the tuning dataset.

With this method, if the search returns a very different idea of
the game than the eval, we will have a better time matching things
up. IE, suppose the eval is drawn, but a tactical line exposes a
clear win. Now, instead of trying to force an otherwise drawn position
to be won via the eval, we will account for it elsewhere.

ELO   | 3.63 +- 2.82 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 4.00]
Games | N: 24380 W: 5238 L: 4983 D: 14159
http://chess.grantnet.us/viewTest/3510/

ELO   | 5.11 +- 3.34 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 4.00]
Games | N: 13120 W: 2169 L: 1976 D: 8975
http://chess.grantnet.us/viewTest/3511/

BENCH : 7,154,736