Scale eval to draw with a lone minor
In >99,9% of positions with a minor against a king and pawns, the side with the minor can at best get a draw, but the raw eval is frequently +1 to +2 for the side with the minor. Scaling this down helps ensuring that a winning line is not discarded in favor of such a draw.

ELO   | 2.53 +- 1.96 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 49225 W: 10216 L: 9857 D: 29152
http://chess.grantnet.us/viewTest/3592/

ELO   | 1.70 +- 1.15 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 3.04 (-2.94, 2.94) [0.00, 5.00]
Games | N: 106150 W: 16420 L: 15901 D: 73829
http://chess.grantnet.us/viewTest/3593/

BENCH : 7,041,633