Advise the kernel to use huge pages for hash
Align the transposition table allocation by 2M to make it huge page
friendly and advise the kernel to use huge pages.

In practice, huge pages with auto-defragment may always be enabled in
the system, in which case this patch is expected to have close to zero
gain. But this depends on the values in
/sys/kernel/mm/transparent_hugepage/enabled and
/sys/kernel/mm/transparent_hugepage/defrag, which are distro-specific.
Not all distros use always/always here, as these settings are not
automatic wins for all applications.

But in any case, it won't hurt to align the TT by the huge page size
and explicitly request huge pages to ensure huge pages when they are
supported.

NO FUNCTIONAL CHANGE.

BENCH 8,120,458