Early game bonus for controlling center squares
The center of the board is especially important because piece maneuvers between different parts of the board often involve crossing through the center. Limited space often means that the pieces get in each other's way and can't coordinate properly. By giving a bonus to each occupied or attacked square that the enemy doesn't attack in the 16 center squares of the board, this patch encourages to take more aggressively control of the center.

This term becomes much less useful if not detrimental in the endgame, when there isn't enough pieces left to contest the control of some squares, so a threshold of minor and major count is used to enable it, while the values goes down to 0 through a negative eg value. Tests without this method failed to provide an elo gain.

The conditions on relevant squares for the bonus, as the values for the bonus and the threshold, are mostly guesswork and could most likely be improved with elo gain.

ELO   | 4.18 +- 3.25 (95%)
SPRT  | 12.0+0.12s @ 1.275Mnps Threads=1 Hash=8MB
LLR   | 3.00 (-2.94, 2.94) [0.00, 5.00]
Games | N: 17720 W: 3696 L: 3483 D: 10541
http://chess.grantnet.us/viewTest/4484/

ELO   | 3.40 +- 2.67 (95%)
SPRT  | 60.0+0.6s @ 1.275Mnps Threads=1 Hash=64MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 5.00]
Games | N: 20353 W: 3295 L: 3096 D: 13962
http://chess.grantnet.us/viewTest/4485/

BENCH : 7,681,536