Save Tablebase hits into the TT with the actual depth of the search
Before we saved with MAX_PLY - 1, with the intention of ensuring that any position in the tree which stumbles onto a saved TB hit would use the information to make cutoff decisions as the depth was the highest it could be.

However, this causes TB entries to eventually overwhelm the TT, as they are low candidates for replacement due to their extreme depth. This patch resolves this, and encourages the search to reexamine TB lines just as the search would reexamine any other line.

Tested Locally with 6-man Syzygy:
12801 - 12664 - 26535 [0.501] 52000

BENCH : 5,508,011