Update Tuner to have an output closer to evaluate.c
It can be annoying to update the formatting and the array subscripts when copying output for large chunks from the Tuner. This patch allows all terms to have custom strings for array subscripts, and also follows the same pattern for when to linebreak when dealing with terms made up of large arrays.

NO FUNCTIONAL CHANGE

BENCH : 5,188,738