Use Syzygy to filter root moves in TB positions
Before, when reaching a TB root, we would simply return a move maintaining the WDL.
Now, we will search as usual, but exclude moves which fail to maintain the ideal WDL.

This behaviour will not apply when running Ethereal with MultiPV > 1 for analysis.
This behaviour will also not apply if the user specifically requests 'go searchmoves ...'

Testing suggests that this patch is a slowdown, due to the trivial overhead for RootNodes.
I am choosing to ignore this minor speed loss, as it should fade with time.

BENCH : 4,555,163