Tune King Safety and modify some interactions of Safety terms
Implement an updated tuner which makes use of AdaGrad to adapt per-parameter and per-phase learning rates. We also adjust the gradient computations to more closely match the true mathematical definitions that apply due to the added nuance of King Safety and Complexity.

A tune using the normal, Linear, weights has shown to be able to gain elo at STC and be near neutral at LTC. A tune of only the Complexity weights as had similar outcomes. This patch is a tune of only Safety, which demonstrates success with the tuning scheme.

In this patch we modified the King Safety function in two ways. First, we removed the KSFriendlyPawns terms. This term was having very large swings in value during the tuning process. From -30 to 30 and back. Additionally, we removed the component which would increase the accumulated attackers weight by a factor of the number of attacking pieces. The idea which was removed is, in my estimation, considered standard in modern King Safety implementations.

A future patch will try to add KSFriendlyPawns back, now that we have found weights for the other King Safety terms. The hope is that we can now use a lower learning rate, and will again have FriendlyPawns gravitate towards a negative score, which acts as a reduction on the danger of the position.

ELO   | 3.86 +- 2.85 (95%)
SPRT  | 12.0+0.12s Threads=1 Hash=8MB
LLR   | 2.98 (-2.94, 2.94) [0.00, 4.00]
Games | N: 20432 W: 3791 L: 3564 D: 13077
http://chess.grantnet.us/test/6972/

ELO   | 3.38 +- 2.42 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 4.00]
Games | N: 20864 W: 2862 L: 2659 D: 15343
http://chess.grantnet.us/test/6973/

BENCH : 4,466,090