Introduce pawn shelter/storm arrays for king safety
The new arrays are independent of the existing big shelter and storm arrays dimensioned on file and rank and adding a fixed value to the evaluation term. The new arrays don't have file-based values.

The core idea is that the quality of the king's pawn cover or the enemy pawn storm is not a fixed advantage or disadvantage. It oftentimes depends on what else the other side has to threaten the king. By plugging the values into king safety, that has a quadratic component, we can better take into account the convergence of threats that is the true recipe for compromised king safety.

This patch increases significantly the size of each pawn-cache entry. It may be better to compute the shelter/storm safety value on each eval call instead, with the added benefit of cleaner code with it done in the main king evaluation function.

Retuning related eval terms might also show benefits.

STC was tested on the commit before the PK-NN introduction :
ELO   | 6.40 +- 4.25 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 3.05 (-2.94, 2.94) [0.00, 5.00]
Games | N: 9072 W: 1694 L: 1527 D: 5851
http://chess.grantnet.us/test/7474/

LTC was rebased and tested against the latest master :
ELO   | 5.32 +- 3.38 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 9404 W: 1165 L: 1021 D: 7218
http://chess.grantnet.us/test/7481/

BENCH : 4,919,336