Update the Network, trained differently and on FRC data
The previous network was trained to output a score, which was, in training, applied to both the MG and EG components of the evaluation. However, in implementation, we were only applying it to the MG in Ethereal.

This new network is trained to expressly produce a score only for the MG, and to have it only be applied to the MG component of the eval. This is done correctly.

Additionally, the Network is trained using an additionally ~9.6 million FRC positions. This was done to increase diversity, and to resolve an issue where Regression Tests failed to show FRC improvements for the previous network.

Testing was done to confirm that the gains for FRC are massive. Testing was done to confirm that any changes to Standard chess are minimal.

Tests for Fischer Random Chess:

ELO   | 18.90 +- 9.10 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 2.97 (-2.94, 2.94) [0.00, 5.00]
Games | N: 2816 W: 785 L: 632 D: 1399
http://chess.grantnet.us/test/7548/

ELO   | 18.95 +- 8.32 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 3.04 (-2.94, 2.94) [0.00, 5.00]
Games | N: 2624 W: 586 L: 443 D: 1595
http://chess.grantnet.us/test/7549/

Tests for Standard Chess:

ELO   | 1.14 +- 2.53 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | -0.93 (-2.94, 2.94) [0.00, 5.00]
Games | N: 29376 W: 6003 L: 5907 D: 17466
http://chess.grantnet.us/test/7547/

ELO   | 1.63 +- 2.27 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 0.25 (-2.94, 2.94) [0.00, 5.00]
Games | N: 33296 W: 6281 L: 6125 D: 20890
http://chess.grantnet.us/test/7546/

BENCH : 4,933,831