Adding new UCI option SkillLevel from 1 to 100
To use SkillLevel, MultiPV will automatically be bumped up to 256 in order to implement the skill logic.
Each move in the MultiPV is given a random cp variance to ensure slight mistakes.
The magnitude of the variance is determined by the following formula, where x is the SkillLevel:

f(x) = 1000**(-(x - 99)/100) + (-6 * x + 594)

If SkillLevel is set to the default value of 100, Ethereal will not enter any of the SkillLevel logic.
Here are estimates of SkillLevel strength for a human player:

90 -> Grandmaster
80 -> Master
70 -> Tournament player
60 -> Strong online player
40 -> Average online player
20 -> Beginner

NO FUNCTIONAL CHANGE

BENCH : 4,063,377