Fix ponderhit being eaten while preparing a search
If ponderhit comes during the time it takes to process the go command string, it is possible for the search thread to set ponder=1, effectively eating the ponder=0 triggered by ponderhit. This resolves that by locking the pondering value behind a mutex, and only allowing its update after the search as begun.

NO FUNCTIONAL CHANGE

BENCH : 3,776,414