Improve speed to report TB Win/Loss
Ethereal has long been missing the trick where you bound the return value of a PvNode by the Syzygy bounds. This patch adds that trick, which allows Ethereal to return TB Win/Loss at the root, instead of cycling through the aspiration window over and over again without ever reporting the TB position.

In the Stockfish mate finding position set, this patch makes it so Ethereal does not find FEWER mates when using TB, which was PREVIOUSLY the case.

BENCH : 3,530,479