Update Pyrrhic; Advise to not look ahead mmap'ed TB files
Follows discussions that resulted from TCEC's use of 7-man TB. This is the only noticable differenc between Stockfish's implementation, and the implementation of many other engines. This patch was pointed out by Aloril, and I found the conversation that lead to it.

Rubi author claims to have found success in this patch, as have others. I am unable to verify that on my own local machines. However, there is no particular downside to this code.

NO FUNCTIONAL CHANGE

BENCH : 3,808,521