Once again use Syzygy to filter out suboptimal WDL moves in TB Root positions
This code was left disabled in a previous patch, in an attempt to fix a bug that occured at TCEC. The version in this patch is equal, and appears fine in both CCC 1+0 events and TCEC's Cup 11 events. There are some bugs talked about in the Stockfish repo concerning accuracy of WDL results regarding the 50-move rule.

NO FUNCTIONAL CHANGE

BENCH : 3,808,521